autocontent-pro/src/publish/youtube/requirements.txt
boto3==1.34.162
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.140.0
